+++ src/xyz/oihana/schema/constants/traits/products/CustomerOffer.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace xyz\oihana\schema\constants\traits\products ;
+
+/**
+ * The enumeration of all the customer offer properties constants.
+ *
+ * @author  Marc Alcaraz (eKameleon)
+ * @package xyz\oihana\schema\constants\traits\products
+ * @since   1.4.0
+ */
+trait CustomerOffer
+{
+    public const string APPLIED_CONDITION = 'appliedCondition' ;
+    public const string CUSTOMER          = 'customer' ;
+}